move swap logic
