+++ smart-contracts/contracts/cl-vault/src/rewards/mod.rs
@@ -1,6 +1,5 @@
 mod distribution;
 mod helpers;
-pub mod swap;
 
 pub use distribution::*;
 pub use helpers::*;
