+++ smart-contracts/contracts/cl-vault/src/vault/mod.rs
@@ -5,5 +5,5 @@ pub mod concentrated_liquidity;
 pub mod exact_deposit;
 pub mod merge;
 pub mod range;
-mod swap;
+pub mod swap;
 pub mod withdraw;
